feat: split app to components
