+++ src/components/CategoriesContainer.js
@@ -1,14 +1,17 @@
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { fetchCategories, setIsSelected } from '../redux';
+import { fetchCategories } from '../redux';
+
+import { CategoryItem } from './CategoryItem';
+import { Button } from './Button';
+import { FilterButton } from './FilterButton';
 import './CategoriesContainer.scss';
 
 export const CategoriesContainer = () => {
   const loading = useSelector((state) => state.categories.loading);
   const error = useSelector((state) => state.categories.error);
   const categories = useSelector((state) => state.categories.categories);
-  let isForbidden = false;
-  let isApproved = false;
+
   // let total = 22;
   // const [approvedCount, setApprovedCount] = useState(0);
   // let forbiddenCount = total - approvedCount;
@@ -21,7 +24,7 @@ export const CategoriesContainer = () => {
     }
 
     return function (a, b) {
-      if (sortOrder == -1) {
+      if (sortOrder === -1) {
         return b[property].localeCompare(a[property]);
       } else {
         return a[property].localeCompare(b[property]);
@@ -38,11 +41,6 @@ export const CategoriesContainer = () => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
-  const selectCategory = (name) => {
-    dispatch(setIsSelected(name));
-    // setApprovedCount(approvedCount + 1);
-  };
-
   return loading ? (
     <h2>Loading...</h2>
   ) : error ? (
@@ -52,40 +50,24 @@ export const CategoriesContainer = () => {
       <div className='verticalScroll'>
         <div className='categoriesContainer'>
           {categories.map((category, index) => (
-            <div
-              className='category'
-              onClick={() => selectCategory(category.name)}
-            >
-              <button className='status-btn'>
-                {category.isSelected ? (
-                  <i className='fa fa-smile-o'></i>
-                ) : (
-                  <i className='fa fa-simplybuilt'></i>
-                )}
-              </button>
-
-              <p key={index}>{category.name}</p>
-              <i className='fa fa-info tooltip'>
-                <span className='tooltiptext'>{category.description}</span>
-              </i>
-            </div>
+            <CategoryItem
+              key={category.name}
+              category={category}
+              index={index}
+            />
           ))}
         </div>
       </div>
 
       <div className='footer'>
         <div className='buttonContainer'>
-          <button onClick={() => (isForbidden = !isForbidden)} className='btn'>
-            <i className='fa fa-simplybuilt'></i>Forbid All
-          </button>
-          <button onClick={() => (isApproved = !isApproved)} className='btn'>
-            <i className='fa fa-smile-o'></i>Approve All
-          </button>
+          <Button forbid={true} />
+          <Button forbid={false} />
         </div>
         <div className='filterContainer'>
           <p>Filters</p>
-          <button className='btn btn-border'> Approved </button>
-          <button className='btn btn-border'> Forbidden</button>
+          <FilterButton filter={true} title='Approved' />
+          <FilterButton filter={false} title='Forbidden' />
         </div>
       </div>
     </div>
