+++ src/components/CategoriesContainer.scss
@@ -137,6 +137,7 @@
   text-align: center;
   padding: 5px 0;
   border-radius: 6px;
+  font-family: 'Open Sans';
 
   /* Position the tooltip text - see examples below! */
   position: absolute;
