+++ src/components/FilterButton.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+export const FilterButton = ({ filter, title }) => {
+  return (
+    <>
+      {filter ? (
+        <button className='btn btn-border'> {title} </button>
+      ) : (
+        <button className='btn btn-border'> {title}</button>
+      )}
+    </>
+  );
+};
